+++ src/schemas/versionedSection.ts
@@ -5,7 +5,9 @@ export const typeDefs = gql`
     "Get all of the VersionedSection for the specified Section ID"
     sectionVersions(sectionId: Int!): [VersionedSection]
     "Search for VersionedSection whose name contains the search term"
-    publishedSections(term: String!): [VersionedSection]
+    publishedSections(term: String!): [VersionedSectionSearchResult]
+    "Get all of the best practice VersionedSection"
+    bestPracticeSections: [VersionedSection]
   }
 
   "Section version type"
@@ -14,6 +16,29 @@ export const typeDefs = gql`
     DRAFT
     "Published - saved state for use when creating DMPs"
     PUBLISHED
+  }
+
+  type VersionedSectionSearchResult {
+    "The unique identifer for the Object"
+    id: Int
+    "The timestamp when the Object was last modifed"
+    modified: String
+    "The timestamp when the Object was created"
+    created: String
+    "The VersionedSection name/title"
+    name: String!
+    "The VersionedSection introduction"
+    introduction: String
+    "The displayOrder of this VersionedSection"
+    displayOrder: Int!
+    "Whether or not this VersionedSection is designated as a 'Best Practice' section"
+    bestPractice: Boolean
+    "The id of the VersionedTemplate that this VersionedSection belongs to"
+    versionedTemplateId: Int
+    "The name of the VersionedTemplate that this VersionedSection belongs to"
+    versionedTemplateName: String
+    "The number of questions associated with this VersionedSection"
+    versionedQuestionCount: Int
   }
 
   "A snapshot of a Section when it became published."