+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Added
+- Added `VersionedSectionSearchResult` to optimize the query for displaying an org's published sections and the bestPractice sections
 - Added `popularFunders` query to the `affiliations` resolver
 - Added `accessLevel` to projectCollaborator and removed `userId`
 - Added new resolvers related to `projectCollaborators`. Also, when project is created, automatically add user as `projectCollaborator` with `access level`= `OWN`
@@ -76,6 +77,7 @@
 - Added models and resolvers for ProjectContributor, ProjectFunder, ProjectOutput and Project
 
 ### Updated
+- Updated `publishedSections` resolver to return the new `VersionedSectionSearchResult` array
 - Format ORCID identifiers consistently, in the `Contributor` and `User` models, the `projectImport` resolver and `orcid` scalar.
 - Changed a number of GraphQL definitions to PascalCase.
 - Fixed projectCollaborators table which had an FKey on the plans table instead of the projects table
@@ -124,6 +126,7 @@
 - added bestPractice flag to the Section
 
 ### Removed
+- Removed duplicate section check from `Section.create` that was just going off of the "name"
 - Dropped the `PlanVersion` table
 - Removed `prepareAPITarget` function.
 - Removed `id` from `Project` model's constructor (already handled in base `MySQLModel`)
@@ -133,7 +136,7 @@
 ### Fixed
 -Was getting `undefined` bestPractice in `updateTemplate` when none was passed in because of the logic on how it was set. Added a check for whether `bestPractice` is defined before setting value. Also, added an update to `createTemplateVersion`, so that errors from the `generateTemplateVersion` will be caught and passed back in graphql response. Previously, when trying to save a DRAFT of a template, the mutation wouldn't return an error to the client, even though the `Save draft` did not successfully complete. [#265]
 - Removed `Copy of` from in front of copied `Section` and `Template` names [#261]
-- Fixed an issue where adding `templateCollaborators` was failing due to the fact that the `userId` field was required. 
+- Fixed an issue where adding `templateCollaborators` was failing due to the fact that the `userId` field was required.
 - Fixed an issue where adding `projectCollaborators` was failing due to the fact that the `userId` field was required. This should not be required to add a new collaborator [#260]
 - When calling `updatePlanContributors`, the resolver should set isPrimaryContact to `false` for all contributors other than the one marked as isPrimary.
 - Fixed an issue where Jest tests failed on Linux due to case-sensitive file paths. The tests passed on macOS because its file system is case-insensitive by default.